app/requirements.txt
gradio==3.41.2
pillow
numpy
